# apps/cms-scaffold/docs/v2/contracts/BW6.openapi.yaml
openapi: 3.0.3
info:
  title: CMS Scaffold API
  description: "Kernel API for CMS Scaffold. Handwritten OpenAPI is the contract source.\n\nPublic reads live under\
    \ `/api/v1/public/**`. Work surfaces do not share\nthat prefix. Session cookie name is `cms_session` (one cookie).\n\
    Published media is not a public static directory.\n\nSurface is resolved from the `Origin` header (see `cms.identity.surface-origins`).\n\
    A cookie request without `Origin` is treated as the Front surface. A bearer request\nwithout `Origin` may name\
    \ its surface with `X-CMS-Surface`.\n\nEvery error response uses `ErrorEnvelope`; `error.code` is one of `ErrorCode`.\n\
    Each operation lists its error codes in its description.\n\nCommon errors (not repeated per operation):\n- 401\
    \ `UNAUTHENTICATED` (no session) or `SESSION_EXPIRED` (expired or revoked session)\n  on every operation that\
    \ is not marked `security: []`.\n- 403 `CSRF_FAILED` on every POST, PUT, PATCH, DELETE sent with a `cms_session`\
    \ cookie\n  when the `X-CSRF-Token` header does not equal the `cms_csrf` cookie.\n- 400 `VALIDATION_FAILED`\
    \ when a path or query parameter has the wrong type\n  (for example a malformed UUID) or a JSON body cannot\
    \ be parsed.\n- 405 `METHOD_NOT_ALLOWED` when the path exists but the HTTP method does not.\n- 415 `MEDIA_TYPE_NOT_SUPPORTED`\
    \ when a JSON operation receives another Content-Type.\n- 500 `INTERNAL_ERROR` for any unexpected server failure.\n\
...
servers:
...
paths:
...
  /api/v1/principals:
    get:
      operationId: listPrincipals
      tags:
      - Principals
      description: 'Requires `manage_principals` on the Admin surface.

        Errors:

        - 403 SURFACE_FORBIDDEN: called from the Front or Back surface.

        - 403 FORBIDDEN: caller lacks `manage_principals`.


        Q-23: role於SQL/store層篩選，回傳total/size是篩選後筆數，仍不是分頁API。400 VALIDATION_FAILED:重複或未知非空role。'
      responses:
        '200':
          description: Principal list without credential hashes (not paginated; page is always 0)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrincipalList'
              example:
                items:
                - &id001
                  id: 10000000-0000-4000-8000-000000000001
                  username: example-admin
                  displayName: 示範管理員
                  email: null
                  status: active
                  roles:
                  - admin
                  lastLoginAt: null
                page: 0
                size: 1
                total: 1
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
        '400':
          $ref: '#/components/responses/Error400'
      parameters:
      - in: query
        name: role
        required: false
        schema:
          type: string
        description: 單一角色code；缺省或空白不篩選。重複或未知非空role回400 VALIDATION_FAILED。
    post:
      operationId: createPrincipal
      tags:
      - Principals
      description: "Requires `manage_principals` on the Admin surface. When `temporaryPassword` is omitted\nthe\
        \ server generates one and returns it once.\nErrors:\n- 400 VALIDATION_FAILED: username does not match `^[a-z0-9._-]{3,32}$`,\
        \ username is taken,\n  `displayName` is longer than 80, `email` is longer than 254 or already used by another\n\
        \  principal (case-insensitive), or the temporary password is shorter than 12 characters or\n  equals the\
        \ username.\n- 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.\n"
      parameters:
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrincipalRequest'
      responses:
        '201':
          description: Principal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedPrincipal'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/assignable:
    get:
      operationId: listAssignablePrincipals
      tags:
      - Principals
      description: 'Principals that entries of `contentType` can be assigned to (G-04): active, not deleted, holding

        `update` on the type on the Back surface (predicates ignored). The caller needs `update` on the type

        on its own surface. `q` is a case-insensitive substring of username or displayName. Order:

        displayName, then username; at most 20. Only `id` and `displayName` are returned.

        Errors:

        - 400 VALIDATION_FAILED: `contentType` missing, unknown or disabled.

        - 403 FORBIDDEN: caller lacks `update` on the type.

        '
      parameters:
      - in: query
        name: contentType
        required: true
        schema:
          type: string
      - in: query
        name: q
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Assignable principals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignablePrincipalList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}:
    get:
      operationId: getPrincipal
      tags:
      - Principals
      description: 'Errors:

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        - 404 PRINCIPAL_NOT_FOUND: no principal with this id.

        '
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    patch:
      operationId: patchPrincipal
      tags:
      - Principals
      description: "Null or absent properties keep their current value; an empty email string clears the email to\
        \ null.\nErrors:\n- 400 VALIDATION_FAILED: `status` is not a PrincipalStatus, `displayName` is longer than\
        \ 80,\n  or `email` is longer than 254 or used by another principal (case-insensitive).\n- 403 LAST_ADMIN:\
        \ the change would leave no active admin.\n- 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.\n- 404\
        \ PRINCIPAL_NOT_FOUND: no principal with this id.\n\n403 SELF_DISABLE_FORBIDDEN:目標是呼叫者且新status=disabled；先於LAST_ADMIN拒絕。"
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchPrincipalRequest'
            example:
              email: ''
      responses:
        '200':
          description: Updated principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
              example: *id001
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}/disable:
    post:
      operationId: disablePrincipal
      tags:
      - Principals
      description: 'Revokes every session of the principal.

        Errors:

        - 403 LAST_ADMIN: the principal is the last active admin.

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        - 404 PRINCIPAL_NOT_FOUND: no principal with this id.


        403 SELF_DISABLE_FORBIDDEN:不允許停用呼叫者自己，先於LAST_ADMIN拒絕。'
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Disabled principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}/unlock:
    post:
      operationId: unlockPrincipal
      tags:
      - Principals
      description: 'Errors:

        - 403 ACCOUNT_DISABLED: a disabled principal cannot be unlocked.

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        - 404 PRINCIPAL_NOT_FOUND: no principal with this id.

        '
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      responses:
        '200':
          description: Unlocked principal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}/roles:
    put:
      operationId: replacePrincipalRoles
      tags:
      - Principals
      description: "Replaces all role assignments. `editor` and `operator` need a non-empty `contentTypeCodes`.\n\
        Errors:\n- 400 VALIDATION_FAILED: missing or duplicate role code, unknown role, or editor/operator\n  without\
        \ an allowlist.\n- 403 LAST_ADMIN: the change would leave no active admin.\n- 403 SURFACE_FORBIDDEN, FORBIDDEN:\
        \ as listPrincipals.\n- 404 PRINCIPAL_NOT_FOUND: no principal with this id.\n\n403 SELF_DEMOTION_FORBIDDEN:目標是呼叫者、目前有admin且新集合沒有admin；先於LAST_ADMIN拒絕。保留admin時可合法修改其他角色。"
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/RoleAssignmentInput'
      responses:
        '204':
          description: Roles replaced
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}/password:
    post:
      operationId: setPrincipalPassword
      tags:
      - Principals
      description: 'Sets a temporary password and revokes every session of the principal. When the body or

        `temporaryPassword` is omitted the server generates one. The password is returned once.

        Errors:

        - 400 VALIDATION_FAILED: the password is shorter than 12 characters or equals the username.

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        - 404 PRINCIPAL_NOT_FOUND: no principal with this id.

        '
      parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporaryPasswordRequest'
      responses:
        '200':
          description: Temporary password set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemporaryPassword'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/principals/{id}/effective-permissions:
    get:
      operationId: effectivePermissions
      tags:
      - Principals
      description: 'Lists anonymous grants plus the grants of every role of the principal.

        Errors:

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        - 404 PRINCIPAL_NOT_FOUND: no principal with this id.

        '
      parameters:
      - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Effective grants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectivePermissionList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/roles:
    get:
      operationId: listRoles
      tags:
      - Principals
      description: 'Errors:

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        '
      responses:
        '200':
          description: Roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
  /api/v1/roles/{code}/permissions:
    get:
      operationId: getRolePermissions
      tags:
      - Principals
      description: 'Errors:

        - 400 VALIDATION_FAILED: unknown role code.

        - 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.

        '
      parameters:
      - $ref: '#/components/parameters/RoleCode'
      responses:
        '200':
          description: Role permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionList'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      operationId: replaceRolePermissions
      tags:
      - Principals
      description: "Replaces all permissions of the role. Omitted `allowedSurfaces` default to\n`[front, back, admin]`\
        \ for `read_published`, `[admin]` for governance actions\n(`manage_types`, `manage_principals`, `manage_settings`,\
        \ `read_audit`) and\n`[back, admin]` otherwise.\nErrors:\n- 400 VALIDATION_FAILED: unknown role, unknown\
        \ action, unknown surface, duplicate\n  permission, malformed predicate, a predicate without `contentType`,\
        \ a predicate whose field is\n  not an enabled string, enum, ref or principal-ref field with index rows\
        \ in that type (02 §4.1),\n  `contentTypeCode` longer than 64, or an empty list for the anonymous role.\n\
        - 403 LAST_ADMIN: the change would leave no active admin.\n- 403 SURFACE_FORBIDDEN, FORBIDDEN: as listPrincipals.\n"
      parameters:
      - $ref: '#/components/parameters/RoleCode'
      - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/PermissionInput'
      responses:
        '204':
          description: Permissions replaced
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '415':
          $ref: '#/components/responses/Error415'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
    RoleCode:
      in: path
      name: code
      required: true
      schema:
        type: string
...
    CsrfHeader:
      in: header
      name: X-CSRF-Token
      required: false
      description: Required when the request carries the cms_session cookie; must equal the cms_csrf cookie.
      schema:
        type: string
  responses:
    Error400:
      description: Bad request. See the operation description for codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error401:
      description: UNAUTHENTICATED, SESSION_EXPIRED or INVALID_CREDENTIALS.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error403:
      description: Forbidden. See the operation description for codes; CSRF_FAILED applies to every unsafe cookie
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error404:
      description: Not found. See the operation description for codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Error415:
      description: Unsupported media type. See the operation description for codes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Error500:
      description: INTERNAL_ERROR.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cms_session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      - requestId
      properties:
        requestId:
          type: string
          description: Echo of X-Request-Id, or a server-generated UUID.
        error:
          type: object
          additionalProperties: false
          required:
          - code
          - message
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: English developer message. Not for display.
            action:
              type: string
              description: Present on FORBIDDEN and SURFACE_FORBIDDEN.
            contentType:
              type: string
              description: Present when the denied action targets a content type.
            surface:
              type: string
              description: Present on FORBIDDEN and SURFACE_FORBIDDEN.
            fields:
              type: array
              minItems: 1
              description: Present on 422 payload, content-type or settings validation errors; every invalid key,
                in validation order.
              items:
                $ref: '#/components/schemas/FieldError'
...
    Surface:
      type: string
      enum:
      - front
      - back
      - admin
    CmsAction:
      type: string
      enum:
      - read_published
      - read_draft
      - create
      - update
      - publish
      - unpublish
      - delete
      - archive
      - manage_media
      - manage_types
      - manage_principals
      - manage_settings
      - read_audit
    PrincipalStatus:
      type: string
      enum:
      - active
      - disabled
      - locked
...
    AssignablePrincipalList:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          maxItems: 20
          items:
            type: object
            additionalProperties: false
            required:
            - id
            - displayName
            properties:
              id:
                type: string
                format: uuid
              displayName:
                type: string
    Principal:
      type: object
      additionalProperties: false
      required:
      - id
      - username
      - displayName
      - email
      - status
      - roles
      - lastLoginAt
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PrincipalStatus'
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
          description: 實際角色 assignments 的去重 code，字典序；沒有 grants 的角色仍列出。
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: 成功登入的實際時間；从未登入為null。
    PrincipalList:
      type: object
      additionalProperties: false
      required:
      - items
      - page
      - size
      - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Principal'
        page:
          type: integer
          description: Always 0 in this version.
        size:
          type: integer
        total:
          type: integer
    CreatePrincipalRequest:
      type: object
      required:
      - username
      properties:
        username:
          type: string
          pattern: ^[a-z0-9._-]{3,32}$
        displayName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        temporaryPassword:
          type: string
          nullable: true
          minLength: 12
    CreatedPrincipal:
      type: object
      additionalProperties: false
      required:
      - id
      - username
      - displayName
      - email
      - status
      - temporaryPassword
      - roles
      - lastLoginAt
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PrincipalStatus'
        temporaryPassword:
          type: string
          description: Returned once. Never logged.
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
          description: 實際角色 assignments 的去重 code，字典序；沒有 grants 的角色仍列出。
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          description: 成功登入的實際時間；从未登入為null。
    PatchPrincipalRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
          description: Q-24：缺省或null保持原值；空字串清除為null。非空字串維持現有驗證，不trim，不更動建立帳號語義。
        status:
          allOf:
          - $ref: '#/components/schemas/PrincipalStatus'
          nullable: true
    RoleAssignmentInput:
      type: object
      required:
      - code
      properties:
        code:
          type: string
          minLength: 1
        contentTypeCodes:
          type: array
          nullable: true
          items:
            type: string
    TemporaryPasswordRequest:
      type: object
      properties:
        temporaryPassword:
          type: string
          nullable: true
          minLength: 12
    TemporaryPassword:
      type: object
      additionalProperties: false
      required:
      - temporaryPassword
      properties:
        temporaryPassword:
          type: string
    EffectivePermission:
      type: object
      additionalProperties: false
      required:
      - role
      - action
      - contentType
      - allowedSurfaces
      - allowlist
      properties:
        role:
          type: string
        action:
          $ref: '#/components/schemas/CmsAction'
        contentType:
          type: string
          description: Empty string for a grant without a content type.
        allowedSurfaces:
          type: array
          items:
            $ref: '#/components/schemas/Surface'
        allowlist:
          type: array
          items:
            type: string
    EffectivePermissionList:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EffectivePermission'
    Role:
      type: object
      additionalProperties: false
      required:
      - code
      - displayName
      - system
      properties:
        code:
          type: string
        displayName:
          type: string
        system:
          type: boolean
    RoleList:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    Permission:
      type: object
      additionalProperties: false
      required:
      - id
      - action
      - contentTypeCode
      - predicateJson
      - allowedSurfaces
      properties:
        id:
          type: string
          format: uuid
        action:
          $ref: '#/components/schemas/CmsAction'
        contentTypeCode:
          type: string
          nullable: true
        predicateJson:
          type: string
          nullable: true
          description: JSON text, for example {"type":"fieldEquals","field":"ownerPrincipalId","value":"$currentPrincipalId"}.
        allowedSurfaces:
          type: array
          items:
            $ref: '#/components/schemas/Surface'
    PermissionList:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    PermissionInput:
      type: object
      required:
      - action
      properties:
        action:
          $ref: '#/components/schemas/CmsAction'
        contentTypeCode:
          type: string
          nullable: true
        predicateJson:
          type: string
          nullable: true
        allowedSurfaces:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Surface'
...
security:
- cookieAuth: []
- bearerAuth: []
...